+++ ui/client/components/graph/Graph.js
@@ -16,6 +16,7 @@ import NodeDetailsModal from "./node-modal/NodeDetailsModal"
 import NodeUtils from "./NodeUtils"
 import {prepareSvg} from "./svg-export/prepareSvg"
 import {drawGraph, directedLayout, isBackgroundObject, createPaper} from "./GraphPartialsInTS"
+import {FocusableDiv} from "./focusable"
 
 export class Graph extends React.Component {
 
@@ -418,7 +419,7 @@ export class Graph extends React.Component {
       <div id="graphContainer" style={{padding: this.props.padding}}>
         {this.props.showNodeDetailsModal ? <NodeDetailsModal/> : null}
         {!_.isEmpty(this.props.edgeToDisplay) ? <EdgeDetailsModal/> : null}
-        <div ref={this.espGraphRef} id={this.props.divId}></div>
+        <FocusableDiv ref={this.espGraphRef} id={this.props.divId}/>
       </div>
     )