+++ ui/client/types/index.ts
@@ -2,3 +2,5 @@ export * from "./node"
 export * from "./groups"
 export * from "./edge"
 export * from "./process"
+export * from "./validation"
+